change read return arguments from int to float64 for simplicity

+++ sound/sound.go
@@ -2,6 +2,6 @@ package sound
 
 // Reader interface allows you to read audio files
 type Reader interface {
-	Read() ([]int, error)
+	Read() ([]float64, error)
 	SampleRate() uint32
 }
